Introduce ExtendedObjectMetadata, holding Writes
Introduced an additional metadata section to Subnet and
NetworkEndpoint, holding a server-side timestamped record of the
latest write to each section of the object.  The timestamps are full
precision, NOT truncated to a whole number of seconds.

Updated ipam and connection agent to use those write stamps instead of
the truncated creation timestamp in the relevant histograms.  Also
introduced a histogram of latency from (a) setting of status by local
connection agent to (b) creation of remote ifc by remote connection
agent.