Shorten Subnet Validator README section
The README section on the Subnet Validator contained more details
than needed, this commits removes such details.

Also, the explanation on why live lists are needed (as opposed to
cache-based ones) is correct but partial: there are other reasons
besides the one which was reported. This commit makes the
explanation more generic.